rename module path
