+++ protocol/gin4r/go.mod
@@ -1,4 +1,4 @@
-module github.com/go-chassis/go-chassis-extension/protocol/ginserver
+module github.com/go-chassis/go-chassis-extension/protocol/gin4r
 
 go 1.15
 
@@ -8,5 +8,4 @@ require (
 	github.com/go-chassis/go-chassis/v2 v2.1.0
 	github.com/go-chassis/openlog v1.1.2
 	github.com/prometheus/client_golang v1.9.0
-	github.com/swaggo/gin-swagger v1.3.0
 )
